refactor: update attendance components imports to use shadcn directory
