+++ features/attendance/components/profile-dialog.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/shadcn/button';
 import {
   Dialog,
   DialogContent,
@@ -6,18 +6,18 @@ import {
   DialogFooter,
   DialogHeader,
   DialogTitle,
-} from '@/components/ui/dialog';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
+} from '@/components/shadcn/dialog';
+import { Input } from '@/components/shadcn/input';
+import { Label } from '@/components/shadcn/label';
 import {
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from '@/components/ui/select';
-import { Separator } from '@/components/ui/separator';
-import { Switch } from '@/components/ui/switch';
+} from '@/components/shadcn/select';
+import { Separator } from '@/components/shadcn/separator';
+import { Switch } from '@/components/shadcn/switch';
 import { getCycleLabel } from '../../../hooks/attendance/use-attendance-settings';
 import type { ShiftTiming, AttendanceProfile } from '@/types/attendance';
